#5 ♻️refactor: Layout Directory
